+++ src/architecture_modules/models/mp_pose.py
@@ -19,15 +19,14 @@ ANGLES = [LEFT_ELBOW, RIGHT_ELBOW,
           LEFT_HIP, RIGHT_HIP,
           LEFT_ANKLE, RIGHT_ANKLE]
 
-LANDMARK_COLUMNS = ["landmark" + str(ldx) + '_' + dim for dim in ['x', 'y', 'z'] for ldx in range(33)]
-ANGLE_COLUMNS = ["left_elbow_roll", "left_elbow_pitch", "left_elbow_yaw",
-                 "right_elbow_roll", "right_elbow_pitch", "right_elbow_yaw",
-                 "left_knee_roll", "left_knee_pitch", "left_knee_yaw",
-                 "right_knee_roll", "right_knee_pitch", "right_knee_yaw",
-                 "left_hip_roll", "left_hip_pitch", "left_hip_yaw",
-                 "right_hip_roll", "right_hip_pitch", "right_hip_yaw",
-                 "left_ankle_roll", "left_ankle_pitch", "left_ankle_yaw",
-                 "right_ankle_roll", "right_ankle_pitch", "right_ankle_yaw"]
+LANDMARK_2D_COLUMNS = ["landmark" + str(ldx) + "_" + dim for ldx in range(33) for dim in ["x", "y"]]
+LANDMARK_3D_COLUMNS = ["landmark" + str(ldx) + "_" + dim for ldx in range(33) for dim in ["x", "y", "z"]]
+ANGLE_2D_COLUMNS = [joint + "_" + dim for joint in (
+                    ["left_elbow", "right_elbow", "left_knee", "right_knee", "left_hip", "right_hip", "left_ankle", "right_angle"])
+                    for dim in ["roll"]]
+ANGLE_3D_COLUMNS = [joint + "_" + dim for joint in (
+                    ["left_elbow", "right_elbow", "left_knee", "right_knee", "left_hip", "right_hip", "left_ankle", "right_angle"])
+                    for dim in ["roll", "pitch", "yaw"]]
 
 
 class PoseModule(ArchiModule):
@@ -37,13 +36,19 @@ class PoseModule(ArchiModule):
     def __init__(self,
                  min_detection_confidence: float,
                  min_tracking_confidence: float,
-                 swap_rb: bool):
+                 swap_rb: bool,
+                 relative_coordinate: bool):
         self.mp_drawing = mp.solutions.drawing_utils
         self.mp_drawing_styles = mp.solutions.drawing_styles
         self.model = mp.solutions.pose.Pose(
             min_detection_confidence=min_detection_confidence,
             min_tracking_confidence=min_tracking_confidence)
         self.swap_rb = swap_rb
+        self.relative_coordinate = relative_coordinate
+        if relative_coordinate:
+            self.n_dims = 3
+        else:
+            self.n_dims = 2
 
     def predict_step(self, batch):
         if self.swap_rb:
@@ -60,36 +65,31 @@ class PoseModule(ArchiModule):
             for frame in tqdm(frames, "Predicting data[" + str(idx) + "]"):
                 # Predict landmarks
                 result = self.predict_step(frame)
-                landmarks = [np.nan for _ in range(self.n_landmarks * self.n_dims)]
-                angle = [np.nan for _ in range(len(ANGLE_COLUMNS))]
 
-                if result.pose_world_landmarks:
-                    for ldx, landmark in enumerate(result.pose_world_landmarks.landmark):
-                        landmarks[ldx * self.n_dims] = landmark.x
-                        landmarks[ldx * self.n_dims + 1] = landmark.y
-                        landmarks[ldx * self.n_dims + 2] = landmark.z
+                # If relative coordinates
+                if self.relative_coordinate:
+                    landmarks, angle = self.__extract_landmarks_and_angles_on_relative_coordinates(result)
+                # If absolute coordinates
+                else:
+                    landmarks, angle = self.__extract_landmarks_and_angles_on_absolute_coordinates(result, frame)
 
-                    # Analyze angle
-                    for adx, ldx in enumerate(ANGLES):
-                        roll, pitch, yaw = self.__calc_angles_on_3d__(
-                            (landmarks[ldx[0] * self.n_dims], landmarks[ldx[0] * self.n_dims + 1], landmarks[ldx[0] * self.n_dims + 2]),
-                            (landmarks[ldx[1] * self.n_dims], landmarks[ldx[1] * self.n_dims + 1], landmarks[ldx[1] * self.n_dims + 2]),
-                            (landmarks[ldx[2] * self.n_dims], landmarks[ldx[2] * self.n_dims + 1], landmarks[ldx[2] * self.n_dims + 2]))
-                        angle[adx * self.n_dims] = roll
-                        angle[adx * self.n_dims + 1] = pitch
-                        angle[adx * self.n_dims + 2] = yaw
-
+                # Add results
                 results.append(result)
                 preds.append(landmarks)
                 angles.append(angle)
 
+            # Visualize results
             vis_frames = self.__visualize__(frames, results)
 
             if logger:
                 name = dataset.getname(idx)
-                logger.save_landmarks(name, preds, LANDMARK_COLUMNS)
+                if self.relative_coordinate:
+                    logger.save_landmarks(name, preds, LANDMARK_3D_COLUMNS)
+                    logger.save_angles(name, angles, ANGLE_3D_COLUMNS)
+                else:
+                    logger.save_landmarks(name, preds, LANDMARK_2D_COLUMNS)
+                    logger.save_angles(name, angles, ANGLE_2D_COLUMNS)
                 logger.save_videos(name, vis_frames, dataset.getmeta(idx))
-                logger.save_angles(name, angles, ANGLE_COLUMNS)
 
     def __visualize__(self, frames, results):
         vis_frames = []
@@ -103,6 +103,49 @@ class PoseModule(ArchiModule):
                     landmark_drawing_spec=self.mp_drawing_styles.get_default_pose_landmarks_style())
             vis_frames.append(frame)
         return vis_frames
+
+    def __extract_landmarks_and_angles_on_relative_coordinates(self, result):
+        landmarks = [np.nan for _ in range(self.n_landmarks * self.n_dims)]
+        angle = [np.nan for _ in range(len(ANGLE_3D_COLUMNS))]
+        # Arrange landmarks
+        for ldx, landmark in enumerate(result.pose_world_landmarks.landmark):
+            landmarks[ldx * self.n_dims] = landmark.x
+            landmarks[ldx * self.n_dims + 1] = landmark.y
+            landmarks[ldx * self.n_dims + 2] = landmark.z
+
+            # Analyze angle
+            for adx, ldx in enumerate(ANGLES):
+                roll, pitch, yaw = self.__calc_angles_on_3d__(
+                    (landmarks[ldx[0] * self.n_dims], landmarks[ldx[0] * self.n_dims + 1],
+                     landmarks[ldx[0] * self.n_dims + 2]),
+                    (landmarks[ldx[1] * self.n_dims], landmarks[ldx[1] * self.n_dims + 1],
+                     landmarks[ldx[1] * self.n_dims + 2]),
+                    (landmarks[ldx[2] * self.n_dims], landmarks[ldx[2] * self.n_dims + 1],
+                     landmarks[ldx[2] * self.n_dims + 2]))
+                angle[adx * self.n_dims] = roll
+                angle[adx * self.n_dims + 1] = pitch
+                angle[adx * self.n_dims + 2] = yaw
+
+        return landmarks, angle
+
+    def __extract_landmarks_and_angles_on_absolute_coordinates(self, result, frame):
+        landmarks = [np.nan for _ in range(self.n_landmarks * self.n_dims)]
+        angle = [np.nan for _ in range(len(ANGLE_2D_COLUMNS))]
+        # Arrange landmarks
+        image_rows, image_cols, _ = frame.shape
+        for ldx, landmark in enumerate(result.pose_landmarks.landmark):
+            landmarks[ldx * self.n_dims] = landmark.x * image_cols
+            landmarks[ldx * self.n_dims + 1] = landmark.y * image_rows
+
+            # Analyze angles
+            for adx, ldx in enumerate(ANGLES):
+                roll = self.__calc_angle_on_2d__(
+                    (landmarks[ldx[0] * self.n_dims], landmarks[ldx[0] * self.n_dims + 1]),
+                    (landmarks[ldx[1] * self.n_dims], landmarks[ldx[1] * self.n_dims + 1]),
+                    (landmarks[ldx[2] * self.n_dims], landmarks[ldx[2] * self.n_dims + 1]))
+                angle[adx] = roll
+
+        return landmarks, angle
 
     def __calc_angles_on_3d__(self, p1, p2, p3):
         angle_xy = self.__calc_angle_on_2d__((p1[0], p1[1]),